Struct reural_network::matrix::Matrix

pub struct Matrix<T> { /* fields omitted */ }

A matrix is a 2-dimensional structure with specific dimensions that can hold data of any type.

Rows and columns of a matrix are zero-indexed, meaning that the top left element of the matrix is in row 0 and column 0, and the bottom right elememt is in row rows - 1 and column columns - 1, where rows and columns are the number of rows and columns the matrix has, respectively.

Example

Create a new matrix from a slice of data, add a scalar value to it and transpose the result:

use std::num::NonZeroUsize;
use reural_network::matrix::Matrix;

let rows = NonZeroUsize::new(2).unwrap();
let columns = NonZeroUsize::new(3).unwrap();
let elements: [f64; 6] = [2.3, 4.0, 3.3, -1.465, 0.0, -42.0];

// Create a 2x3 matrix.
let matrix = Matrix::from_slice(rows, columns, &elements).unwrap();

// Add 7.3 to each element.
let sum = &matrix + 7.3;

// Transpose the sum to get a 3x2 matrix.
let transposed = sum.transpose();

assert_eq!(transposed.as_slice(), &[9.6, 5.835, 11.3, 7.3, 10.6, -34.7]);

Size Limits

A matrix can hold a maximum number of ::std::usize::MAX elements. When creating a new matrix, it is checked if the given number of rows and columns would create a matrix that would exceed this size limit. In this case, the matrix cannot be created.

Supported Mathematical Operations

The following mathematical operations are supported for matrices Matrix<T>:

Addition*

Element-wise addition of two matrices with the same dimensions.

Scalar addition of a matrix and a scalar value.

Subtraction*

Element-wise subtraction of two matrices with the same dimensions.

Scalar subtraction of a scalar value from a matrix.

Multiplication*

Element-wise multiplication (Hadamard product) of two matrices with the same dimensions.

Scalar multiplication of a scalar value to a matrix.

Division*

Element-wise division of two matrices with the same dimensions.

Scalar division of a matrix by a scalar value.

Bitwise AND*

Element-wise bitwise AND of two matrices with the same dimensions.

Scalar bitwise AND of a matrix and a scalar value.

Bitwise OR*

Element-wise bitwise OR of two matrices with the same dimensions.

Scalar bitwise OR of a matrix and a scalar value.

Bitwise XOR*

Element-wise bitwise XOR of two matrices with the same dimensions.

Scalar bitwise XOR of a matrix and a scalar value.

Left Shift*

Element-wise left shift of a matrix by another matrix with the same dimensions.

Scalar left shift of a matrix by a scalar value.

Right Shift*

Element-wise right shift of a matrix by another matrix with the same dimensions.

Scalar right shift of a matrix by a scalar value.

Negation:* Negate all elements of a matrix.

Logical Negation:* Logically negate all elements in a matrix.

Matrix Multiplication: compute the matrix product of two matrices (matrix_mul).

Transposition: flipping a matrix over its diagonal (transpose).

Map: change each element in a matrix based on a closure (map).

These operations use a naive implementation without any considerations for performance.

The binary operations are also implemented as assignment operators for scalar values (e.g. matrix += 4.5). However, they are not implemented as assignment operators for matrices (e.g. matrix += other_matrix) since both matrices must have the same dimensions. If the other matrix did not have the correct dimensions, we could either not perform the operation on matrix, or we could panic. Since both of these possibilities can be very surprising to the user, the assignment operators are not implemeneted for matrices. Instead, you will have to resort to solutions like:

let result = matrix + other_matrix;
let matrix = result.unwrap();

* The operation must be implemeneted for the type T.

Implementations

impl<T> Matrix<T>

pub fn as_slice(&self) -> &[T]

Get the data of the matrix as a 1-dimensional slice.

For a matrix with m rows and n columns, the first row of the matrix will become the first m elements in the slice, the second row will become the second m elements and so on.

Example

The matrix

[0 1 2]
[3 4 5]

will produce the following slice:

let data: &[usize] = &[0, 1, 2, 3, 4, 5];

pub fn get_number_of_columns(&self) -> usize

Get the number of columns in the matrix.

pub fn get_number_of_rows(&self) -> usize

Get the number of rows in the matrix.

pub fn map_ref_mut<F>(&mut self, mapping: F) where
F: FnMut(&mut T, usize, usize),

Mutate each element in the matrix in place as given by the closure mapping.

The mapping closure has three parameters, in this order:

The value of the current element as a mutable reference.

The row of the current element.

The column of the current element.

Example

Convert a matrix of temperatures in °C to K:

use std::num::NonZeroUsize;
use reural_network::matrix::Matrix;

let rows: NonZeroUsize = NonZeroUsize::new(2).unwrap();
let columns: NonZeroUsize = NonZeroUsize::new(3).unwrap();
let temperatures: [f64; 6] = [0.0, 10.0, 25.0, 50.0, 75.0, 100.0];
let mut matrix: Matrix<f64> = Matrix::from_slice(rows, columns, &temperatures).unwrap();

// Convert Celsius to Kelvin.
matrix.map_ref_mut(|celsius, _row, _column| *celsius += 273.15);
assert_eq!(matrix.as_slice(), [273.15, 283.15, 298.15, 323.15, 348.15, 373.15]);

impl<T> Matrix<T> where
T: Copy,

pub fn new(
rows: NonZeroUsize,
columns: NonZeroUsize,
default: T
) -> Result<Matrix<T>>

Create a new matrix with the given dimensions and the given default value in all elements.

The product of the number of rows and the number of columns must not exceed the maximum usize value, ::std::usize::MAX. Otherwise, an Error::DimensionsTooLarge will be returned.

Example

A 2x3 matrix with a default value of 0.25 for all elements can be created with the following lines of code:

use std::num::NonZeroUsize;
use reural_network::matrix::Matrix;

let rows = NonZeroUsize::new(2).unwrap();
let columns = NonZeroUsize::new(3).unwrap();
let matrix: Matrix<f64> = Matrix::new(rows, columns, 0.25).unwrap();

pub fn from_slice(
rows: NonZeroUsize,
columns: NonZeroUsize,
data: &[T]
) -> Result<Matrix<T>>

Convert a slice into a matrix of the given dimensions.

For a matrix with m rows and n columns, the first m elements in the slice will become the first row in the matrix, the second m elements will become the second row and so on.

The product of the number of rows and the number of columns must not exceed the maximum usize value, ::std::usize::MAX. If it does, an Error::DimensionsTooLarge will be returned. Furthermore, the product must be equal to the length of the given data slice. Otherwise, an Error::DimensionMismatch will be returned.

Example

A 2x3 matrix can be created from a slice of length 6 with the following lines of code:

use std::num::NonZeroUsize;
use reural_network::matrix::Matrix;

let rows = NonZeroUsize::new(2).unwrap();
let columns = NonZeroUsize::new(3).unwrap();
let matrix: Matrix<i32> = Matrix::from_slice(rows, columns, &[0, 1, 2, 3, 4, 5]).unwrap();

This will produce the matrix:

[0 1 2]
[3 4 5]

pub fn get(&self, row: usize, column: usize) -> Result<T>

Get the value in the given row and column.

If the row or column value is larger than the number of rows or columns in the matrix, respectively, an Error::CellOutOfBounds will be returned.

If it can be guaranteed that the row and column values do not exceed the dimensions of the matrix, you can also use get_unchecked.

pub unsafe fn get_unchecked(&self, row: usize, column: usize) -> T

Get the value in the given row and column.

This method does not check if the row and column parameters have valid values. If they are greater than or equal to the dimensions of the matrix, the resulting index will be out of bounds and the call will panic.

This method should only be used if it is guaranteed by the caller that the row and column are within the dimensions of the matrix. If this cannot be guaranteed, use get instead.

Safety

If the row or column are out of bounds of this matrix, an invalid index in the internal data structure will be accessed. This will cause the method to panic.

pub fn map<F>(&mut self, mapping: F) where
F: Fn(T, usize, usize) -> T,

Map each element in the matrix to a new element as given by the closure mapping.

The mapping closure has three parameters, in this order:

The value of the current element.

The row of the current element.

The column of the current element.

It must return the new value of the current element.

Example

Convert a matrix of temperatures in °C to °F:

use std::num::NonZeroUsize;
use reural_network::matrix::Matrix;

let rows: NonZeroUsize = NonZeroUsize::new(2).unwrap();
let columns: NonZeroUsize = NonZeroUsize::new(3).unwrap();
let temperatures: [usize; 6] = [0, 10, 25, 50, 75, 100];
let mut matrix: Matrix<usize> = Matrix::from_slice(rows, columns, &temperatures).unwrap();

// Convert Celsius to Fahrenheit (these values come out as perfect integers).
matrix.map(|celsius, _row, _column| (celsius * 9 / 5) + 32);
assert_eq!(matrix.as_slice(), [32, 50, 77, 122, 167, 212]);

pub fn transpose(&self) -> Matrix<T>

Transpose this matrix.

Example

A 2x3 matrix

[0 1 2]
[3 4 5]

will become a 3x2 matrix:

[0 3]
[1 4]
[2 5]

In code, this will look as follows:

use std::num::NonZeroUsize;
use reural_network::matrix::Matrix;

let rows: NonZeroUsize = NonZeroUsize::new(2).unwrap();
let columns: NonZeroUsize = NonZeroUsize::new(3).unwrap();
let matrix: Matrix<usize> = Matrix::from_slice(rows, columns, &[0, 1, 2, 3, 4, 5]).unwrap();

let transposed: Matrix<usize> = matrix.transpose();
assert_eq!(transposed.get_number_of_rows(), 3);
assert_eq!(transposed.get_number_of_columns(), 2);
assert_eq!(transposed.as_slice(), &[0, 3, 1, 4, 2, 5]);

impl<T> Matrix<T> where
T: Add<T, Output = T> + Mul<T, Output = T> + Copy,

pub fn matrix_mul(&self, other: &Matrix<T>) -> Result<Matrix<T>>

Compute the matrix product of self and other and return the result.

The number of columns in self must be equal to the number of rows in other. Otherwise, Error::DimensionMismatch will be returned.

The resulting matrix will have the dimensions self.rows x other.columns. If these dimensions would exceed the maximum size of matrices, Error::DimensionsTooLarge will be returned.

Example

// `m1` (2x3):
// [1   2   3]
// [4   5   6]
let rows_m1 = NonZeroUsize::new(2).unwrap();
let columns_m1 = NonZeroUsize::new(3).unwrap();
let data_m1: [usize; 6] = [1, 2, 3, 4, 5, 6];
let m1: Matrix<usize> = Matrix::from_slice(rows_m1, columns_m1, &data_m1).unwrap();

// `m2` (3x2):
// [7    8 ]
// [9    10]
// [11   12]
let rows_m2 = NonZeroUsize::new(3).unwrap();
let columns_m2 = NonZeroUsize::new(2).unwrap();
let data_m2: [usize; 6] = [7, 8, 9, 10, 11, 12];
let m2: Matrix<usize> = Matrix::from_slice(rows_m2, columns_m2, &data_m2).unwrap();

// Calculate `m3` as the matrix product of `m1` and `m2`.
// `m3` (2x2):
// [58    64 ]
// [139   154]
let m3: Matrix<usize> = m1.matrix_mul(&m2).unwrap();
assert_eq!(m3.get_number_of_rows(), 2);
assert_eq!(m3.get_number_of_columns(), 2);
assert_eq!(m3.as_slice(), &[58, 64, 139, 154]);

impl Matrix<f64>

pub fn from_random(
rows: NonZeroUsize,
columns: NonZeroUsize
) -> Result<Matrix<f64>>

Create a new matrix with the given dimensions and random elements in the inclusive range [0.0, 1.0] (i.e., including both 0.0 and 1.0).

The product of the number of rows and the number of columns must not exceed the maximum usize value, ::std::usize::MAX. Otherwise, an Error::DimensionsTooLarge will be returned.

Example

A 2x3 matrix with a default value of 0.25 for all elements can be created with the following lines of code:

use std::num::NonZeroUsize;
use reural_network::matrix::Matrix;

let rows = NonZeroUsize::new(2).unwrap();
let columns = NonZeroUsize::new(3).unwrap();
let matrix: Matrix<f64> = Matrix::from_random(rows, columns).unwrap();